35,70 test and bagfix

+++ include/tbitfield.h
@@ -12,7 +12,7 @@
 
 using namespace std;
 
-typedef unsigned int TELEM;
+typedef unsigned long long TELEM;
 
 class TBitField
 {
@@ -44,7 +44,8 @@ public:
   TBitField  operator~(void);                // отрицание                  (#С)
 
   friend istream &operator>>(istream &istr, TBitField &bf);       //      (#О7)
-  friend ostream &operator<<(ostream &ostr, const TBitField &bf); //      (#П4)
+  friend ostream &operator<<(ostream &ostr, const TBitField &bf);
+  //      (#П4)
 };
 // Структура хранения битового поля
 //   бит.поле - набор битов с номерами от 0 до BitLen
